backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2
numpy
requests
pandas
yfinance
datetime
kafka-python
pyspark
confluent-kafka
lxml
